Compiler warning produced by gcc 4.4 on Linux Ubuntu 10 x86_64
git-svn-id: file:///home/michael/svn/yuni/trunk@1046 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/directory/make.cpp
@@ -18,7 +18,7 @@ namespace Directory
 
 	# ifdef YUNI_OS_WINDOWS
 
-	bool WindowsMake(const char* path, unsigned int len)
+	bool WindowsMake(const char* path, size_t len)
 	{
 		if (len)
 		{
@@ -75,7 +75,7 @@ namespace Directory
 
 	# else
 
-	bool UnixMake(const char* path, unsigned int len, unsigned int mode)
+	bool UnixMake(const char* path, size_t len, unsigned int mode)
 	{
 		if (!len || NULL == path || '\0' == *path)
 			return true;
